--- packages/forma-36-tokens/src/tokens/colors/colors-green-deprecated.js
@@ -1,11 +1,0 @@
-const scaleGreen = require('./scale-green');
-
-const colorsGreen = {
-  'color-green-dark': scaleGreen['green-700'],
-  'color-green-base': scaleGreen['green-600'],
-  'color-green-mid': scaleGreen['green-500'],
-  'color-green-light': scaleGreen['green-300'],
-  'color-green-lightest': scaleGreen['green-100'],
-};
-
-module.exports = colorsGreen;
