--- packages/forma-36-tokens/src/tokens/colors/scale-green.js
@@ -1,13 +1,0 @@
-const scaleGreen = {
-  'green-100': '#EAF9E8',
-  'green-200': '#CDF3C6',
-  'green-300': '#9ED696',
-  'green-400': '#5DB057',
-  'green-500': '#008539',
-  'green-600': '#006D23',
-  'green-700': '#00550C',
-  'green-800': '#004700',
-  'green-900': '#003F00',
-};
-
-module.exports = scaleGreen;
